Major refactor of the latitude subsystem of Ellipsoid
In order to better accomodate implementation of the new series
expansions for auxiliary latitudes recently published by Charles
Karney, we introduce the AuxLatitudeFourierCoefficients data
structure, and its companions AuxLatitudeCoefficients.

The PROJ implementation of the same functionality relies on bare
array pointers, which we could as well. But the approach taken is
more in line with Rust's type safety focus. Also it mitigates the
risk of (accidental or adversarial) provision of too short arrays,
which potentially opens up for stack smashing.