Clean up inner_op tests and return types
This commit switches the return type of the inner_op fwd/inv
methods, from `Result<usize, Error>` to a plain `usize`.

As part of the mopping-up of collateral effects, test cases
using the older PROJ style direct calls have been switched
to indirect calling through a context object.

Rationale
---------

It is fully meaningful for the operator-constructors to return
a `Result<Op, Error>`, since they can describe exactly what kind
of error lead to the failure to construct.

Having the inner_op `fwd/inv`-methods returning a `Result<usize, Error>`,
on the other hand, was not very meaningful, since inner_ops operate
on a number of coordinates at a time and hence transient errors may
be of any multitude and of any kind.

The only measure the inner_op can meaningfully return is the number
of coordinates succesfully operated on. Errors will typically result
in NANs, which clearly and unmistakably conveys the message that "in
this case something went wrong". The exact nature of the individual
errors is mostly of interest in debugging, and can be connumicated
by logging.

While changing the `Result<usize,Error>` return type to a plain usize,
a large number of tests also needed adaptation. In cases where the
tests used the older `op.apply(ctx...)` calling convention, this has
been switched to the newer `ctx.apply(op...)`, hence introducing a
bit more consistency in the test material.