More coordinate mopping
The coordinate trait has been removed: The implemntation of its
associated methods have been changed to be a convention for
implemnetation of CoordinateTuples of any kind.

This allows us to just have the number of initializers needed for
the dimensionality of the given coordinate tuple type, rather than
using 4 dimensions for 2D data.

Also, this commit adds the missing implementation of the
CoordinateSet trait for a number of Coor32 containers.