Extend the 'plain' context provider and use it!
The 'plain' context makes it possible to work with external
grids and other resources. This commit extends the 'plain'
functionality and tests its practicality by using it in the
'kp' coordinate processing program.

More test material has been added: A collection of transformations
and a deformation grid, to support the testing of the 'plain'
functionality.