Simplify interpolation in stacks of grids
Introduce the 'grids_at' function, which tries to interopolate in
a stack of grids, and returns the first succesful interpolation
(or 'None' in case of failure).

Prior to this, the gridshift operator needed to explicitly loop
over all grids in the stack, in order to find the one containing
the interpoland. This made especially the inverse case of datum
shifts, where iteration is needed, needlessly complex.