Better format for macro arguments and defaults
Macro definitions now use the syntax 'macro:name arg=(0)' to provide
a default value (here 0) if the argument 'arg' is not given at
instantiation time.

Named arguments can now also provide a default value using the
syntax 'macro:name arg=$foo(1)' to indicate that 'arg' should
be set to the value of 'foo', and if 'foo' is not given, to '1'.