Make Op.steps an Option<Vec<op>>
Until now, Op.steps was initialized as an empty vector
for all instantiations of Op, even though the steps
element only makes sens for pipeline Ops.

This commit changes Op.steps to an option, and introduces
the Op::is_pipeline() method. Other than that, this commit
is just an implementation detail.