Provide a meaningful `_name` for pipelines
Pipelines do not really use the `params: ParsedParameters` field.
As a consequence, it was not noted until now, that the `_name`
item was given the meaningless value of "everything up until the
first whitespace in the instantiated_as definition".

This commit changes that, such that `_name` is now the entire
definition