Preparation for unigrid
A messy commit, cleaning up here and there, preparing for the unigrid
construction, where all grids are stored externally: In the initial
case in a single concatenated file, but potentially also in databases
or by memory mapping the concatenation.

One of the cleanups are not directly related to the prep work, but
just conceptual simplifications, prior to this major change: The
elmination of the unfortunate overload of the adjective "Plain".

Most prominently, Plain is the name of the Plain constructor. It
still is. But the two other "plain cases", i.e. the constructor for
`BaseGrid`, and the constructor helper for `Op`erations, have been
renamed to `new(...)`and `basic(...)`, respectively.

More directly related to the unigrid idea, the `Context` trait is exteded
with the `get_grid_value(...)` method, which accesses a specific, indexed,
value of a correction grid. The default implementation returns an
`Error::Unsupported` message.

Also, the signature for the `Grid::at(...)` method has been extended with
an `Option<&dyn Context>` arg, for passing on the context which should
provide the lookup for the unigrid index value.