Make BaseGrid feature compatible with Ntv2Grid
There is much overlap between BaseGrid and Ntv2Grid. Essentially BaseGrid
just needs a vector of subgrids to be able to fulfill Ntv2Grid's needs.

Since Ntv2Grid's multi-toplevel-grid feature was untested, and since I have
never seen a NTv2 file with more than 1 toplevel grid, the multi-toplevel-grid
feature has not been implemented for the BaseGrid NTv2 reader.

Also the multilevel hierarchical grid search has been eliminated, and
replaced by sorting of the subgrids into an order, that should give the
same results from a plain linear search.

This will, in some cases, mean that more subgrids must be checked for a grid
search. But as each check is simpler, and avoids allocations, this may show
to be faster.

In preparation for the removal of the Ntv2Grid struct, this commit also
reimplements all existing tests for the BaseGrid implementation, and runs
them, with identical results, in tandem with their Ntv2Grid counterparts.